Kan 168/test

+++ app/domains/action/models.py
@@ -1,5 +1,5 @@
 # app\domains\action\models.py
-from sqlalchemy import Column, BigInteger, String, Enum, DateTime, Boolean, ForeignKey, Text, Integer, Date, func
+from sqlalchemy import Column, String, Enum, DateTime, Boolean, ForeignKey, Text, Integer, Date, func
 from app.infra.database.base import Base
 import enum
 
@@ -27,10 +27,10 @@ class ReportFormat(str, enum.Enum):
 class ActionItem(Base):
     __tablename__ = "action_items"
 
-    id           = Column(BigInteger, primary_key=True, autoincrement=True)
-    meeting_id   = Column(BigInteger, ForeignKey("meetings.id"), nullable=False)
+    id           = Column(Integer, primary_key=True, autoincrement=True)
+    meeting_id   = Column(Integer, ForeignKey("meetings.id"), nullable=False)
     content      = Column(Text, nullable=False)
-    assignee_id  = Column(BigInteger, ForeignKey("users.id"), nullable=True)
+    assignee_id  = Column(Integer, ForeignKey("users.id"), nullable=True)
     due_date     = Column(Date, nullable=True)
     status       = Column(Enum(ActionStatus), default=ActionStatus.pending)
     detected_at  = Column(DateTime, nullable=False)
@@ -40,8 +40,8 @@ class ActionItem(Base):
 class WbsEpic(Base):
     __tablename__ = "wbs_epics"
 
-    id           = Column(BigInteger, primary_key=True, autoincrement=True)
-    meeting_id   = Column(BigInteger, ForeignKey("meetings.id"), nullable=False)
+    id           = Column(Integer, primary_key=True, autoincrement=True)
+    meeting_id   = Column(Integer, ForeignKey("meetings.id"), nullable=False)
     title        = Column(String(200), nullable=False)
     order_index  = Column(Integer, nullable=False)
     jira_epic_id = Column(String(100), nullable=True)
@@ -50,10 +50,10 @@ class WbsEpic(Base):
 class WbsTask(Base):
     __tablename__ = "wbs_tasks"
 
-    id             = Column(BigInteger, primary_key=True, autoincrement=True)
-    epic_id        = Column(BigInteger, ForeignKey("wbs_epics.id"), nullable=False)
+    id             = Column(Integer, primary_key=True, autoincrement=True)
+    epic_id        = Column(Integer, ForeignKey("wbs_epics.id"), nullable=False)
     title          = Column(String(200), nullable=False)
-    assignee_id    = Column(BigInteger, ForeignKey("users.id"), nullable=True)
+    assignee_id    = Column(Integer, ForeignKey("users.id"), nullable=True)
     priority       = Column(Enum(Priority), default=Priority.medium)
     due_date       = Column(Date, nullable=True)
     progress       = Column(Integer, default=0)
@@ -67,9 +67,9 @@ class WbsTask(Base):
 class Report(Base):
     __tablename__ = "reports"
 
-    id         = Column(BigInteger, primary_key=True, autoincrement=True)
-    meeting_id = Column(BigInteger, ForeignKey("meetings.id"), nullable=False)
-    created_by = Column(BigInteger, ForeignKey("users.id"), nullable=False)
+    id         = Column(Integer, primary_key=True, autoincrement=True)
+    meeting_id = Column(Integer, ForeignKey("meetings.id"), nullable=False)
+    created_by = Column(Integer, ForeignKey("users.id"), nullable=False)
     format     = Column(Enum(ReportFormat), nullable=False)
     file_url   = Column(String(500), nullable=True)
     created_at = Column(DateTime, default=func.now(), nullable=False)
