+++ app/domains/meeting/models.py
@@ -1,5 +1,5 @@
 # app\domains\meeting\models.py
-from sqlalchemy import Column, BigInteger, String, Enum, DateTime, Boolean, ForeignKey, Integer, func
+from sqlalchemy import Column, String, Enum, DateTime, Boolean, ForeignKey, Integer, func
 from app.infra.database.base import Base
 import enum
 
@@ -14,9 +14,9 @@ class DiarizationMethod(str, enum.Enum):
 
 class Meeting(Base):
     __tablename__ = "meetings"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    workspace_id = Column(BigInteger, ForeignKey("workspaces.id"), nullable=False)
-    created_by = Column(BigInteger, ForeignKey("users.id"), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    workspace_id = Column(Integer, ForeignKey("workspaces.id"), nullable=False)
+    created_by = Column(Integer, ForeignKey("users.id"), nullable=False)
     title = Column(String(200), nullable=False)
     meeting_type = Column(String(100), nullable=True)
     status = Column(Enum(MeetingStatus), default=MeetingStatus.scheduled)
@@ -32,28 +32,28 @@ class Meeting(Base):
 class MeetingParticipant(Base):
     __tablename__ = "meeting_participants"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    meeting_id = Column(BigInteger, ForeignKey("meetings.id"), nullable=False)
-    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    meeting_id = Column(Integer, ForeignKey("meetings.id"), nullable=False)
+    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     speaker_label = Column(String(20), nullable=True)
     is_host = Column(Boolean, default=False)
 
 class Agenda(Base):
     __tablename__ = "agendas"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    meeting_id = Column(BigInteger, ForeignKey("meetings.id"), nullable=False)
-    created_by = Column(BigInteger, ForeignKey("users.id"), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    meeting_id = Column(Integer, ForeignKey("meetings.id"), nullable=False)
+    created_by = Column(Integer, ForeignKey("users.id"), nullable=False)
     
     created_at = Column(DateTime, default=func.now(), nullable=False)
 
 class AgendaItem(Base):
     __tablename__ = "agenda_items"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    agenda_id = Column(BigInteger, ForeignKey("agendas.id"), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    agenda_id = Column(Integer, ForeignKey("agendas.id"), nullable=False)
     title = Column(String(200), nullable=False)
-    presenter_id = Column(BigInteger, ForeignKey("users.id"), nullable=True)
+    presenter_id = Column(Integer, ForeignKey("users.id"), nullable=True)
     estimated_minutes = Column(Integer, nullable=True)
     reference_url = Column(String(500), nullable=True)
     order_index = Column(Integer, nullable=False)
@@ -61,9 +61,9 @@ class AgendaItem(Base):
 class SpeakerProfile(Base):
     __tablename__ = "speaker_profiles"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
-    workspace_id = Column(BigInteger, ForeignKey("workspaces.id"), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
+    workspace_id = Column(Integer, ForeignKey("workspaces.id"), nullable=False)
     voice_model_path = Column(String(500), nullable=True)
     diarization_method = Column(Enum(DiarizationMethod), nullable=False)
     is_verified = Column(Boolean, default=False)
